refactor: move lib/gui/validators => apps/deskflow-gui/validators
