+++ src/apps/deskflow-gui/validators/ComputerNameValidator.cpp
