+++ src/apps/deskflow-gui/validators/ComputerNameValidator.h
