+++ src/apps/deskflow-gui/validators/EmptyStringValidator.cpp
