+++ src/apps/deskflow-gui/validators/EmptyStringValidator.h
