+++ src/apps/deskflow-gui/validators/IStringValidator.cpp
