+++ src/apps/deskflow-gui/validators/LineEditValidator.cpp
