+++ src/apps/deskflow-gui/validators/LineEditValidator.h
