+++ src/apps/deskflow-gui/validators/ScreenDuplicationsValidator.cpp
