+++ src/apps/deskflow-gui/validators/ScreenDuplicationsValidator.h
