+++ src/apps/deskflow-gui/validators/SpacesValidator.cpp
